# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/schools": {
      "get": {
        "operationId": "SchoolsController_findAll",
        "summary": "Get list of all partner schools for dropdown pickers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Alphabetically sorted list of schools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchoolResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Lookups"
        ],
        "security": [
          {
            "jwt": []
          },
          {
            "jwt": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "International Student Application Tracker API",
    "description": "REST API and Grounded AI Diagnostic Agent for International Student Admissions & Partner Management",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "HTTP-only Session JWT Cookie"
      },
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Bearer JWT token authorization"
      }
    },
    "schemas": {
...
      "SchoolResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "School UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "School Name",
            "example": "University of Toronto"
          },
          "country": {
            "type": "string",
            "description": "School Country",
            "example": "Canada"
          }
        },
        "required": [
          "id",
          "name",
          "country"
        ]
      },
...
    }
  }
}